requirements.txt
asgiref==3.9.1
Django==5.2.4
django-environ==0.12.0
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.5.3
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
